requirements.txt
# Core
prefect==2.14.2
pandas==2.0.3
numpy==1.24.3
pyarrow==13.0.0

# Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# File watching
watchdog==3.0.0

# Validation
great-expectations==0.17.19
pyyaml==6.0.1

# Environment
python-dotenv==1.0.0